wallet: move deprecated code in `deprecated.go`
A missing step from the previous utxo manager and tx creator refactor.
